Remove "Trojan.U.NjRat.tr"
I removed the blank line from the end of the file. This procedure is to change the checksum of the file so that it is not recognized as "Trojan.U.NjRat.tr" by GridinSoft and VirusTotal. The change does not change the content of the license but removes the problem when distributing software that contains edge_tts.